requirements.txt
discord.py>=1.1.0
jishaku>=1.12.0.116